Isolate statement insurer in its helper
