+++ lib/safe-pg-migrations/base.rb
@@ -5,6 +5,7 @@ require 'safe-pg-migrations/helpers/logger'
 require 'safe-pg-migrations/helpers/satisfied_helper'
 require 'safe-pg-migrations/helpers/index_helper'
 require 'safe-pg-migrations/helpers/batch_over'
+require 'safe-pg-migrations/helpers/timeout_management'
 require 'safe-pg-migrations/plugins/verbose_sql_logger'
 require 'safe-pg-migrations/plugins/blocking_activity_logger'
 require 'safe-pg-migrations/plugins/statement_insurer/add_column'
